laptop_backend/requirements.txt
flask==3.0.2
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
flask-cors==4.0.0
gunicorn==21.2.0
markdown2==2.4.12

# AI Dependencies - Google Gemini API
google-generativeai>=0.3.0

# Utility Dependencies
numpy>=1.24.0           # Numerical operations
requests>=2.31.0        # HTTP requests
